+++ docs/src/components/Layout/LogoLink.tsx
@@ -1,9 +1,5 @@
 import NextLink from 'next/link';
 
-import { VisuallyHidden } from '@aws-amplify/ui-react';
-
-import { Logo } from '../Logo';
-
 interface LogoLinkProps {
   platform: string;
   onClick?: () => void;
@@ -13,8 +9,7 @@ export const LogoLink = ({ platform, onClick }: LogoLinkProps) => {
   return (
     <NextLink href={`/${platform}`} passHref legacyBehavior>
       <a onClick={onClick} className="docs-logo-link">
-        <VisuallyHidden>Amplify UI Home</VisuallyHidden>
-        <Logo />
+        Amplify UI
       </a>
     </NextLink>
   );