+++ packages/react-geo/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @aws-amplify/ui-react-geo
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aws-amplify/ui-react-core@3.0.2
+
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aws-amplify/ui-react-core@3.0.1
+
 ## 2.0.0
 
 ### Major Changes