+++ packages/react-native-auth/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @aws-amplify/ui-react-native-auth
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [[`4ff26e27e`](https://github.com/aws-amplify/amplify-ui/commit/4ff26e27e4bc8909cc2b86c738eca5085b2a42d1), [`bb141a719`](https://github.com/aws-amplify/amplify-ui/commit/bb141a719fd9bc2d7680e539f2ff047deb88ee7e)]:
+  - @aws-amplify/ui-react-native@2.0.3
+  - @aws-amplify/ui@6.0.2
+  - @aws-amplify/ui-react-core-auth@0.0.4
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [[`9dd0e58e5`](https://github.com/aws-amplify/amplify-ui/commit/9dd0e58e5167d307c2154b3280de3c0e52f607e1)]:
+  - @aws-amplify/ui@6.0.1
+  - @aws-amplify/ui-react-core-auth@0.0.3
+  - @aws-amplify/ui-react-native@2.0.2
+
 ## 0.0.3
 
 ### Patch Changes