+++ packages/react-storage/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @aws-amplify/ui-react-storage
 
+## 3.0.4
+
+### Patch Changes
+
+- Updated dependencies [[`4ff26e27e`](https://github.com/aws-amplify/amplify-ui/commit/4ff26e27e4bc8909cc2b86c738eca5085b2a42d1), [`bb141a719`](https://github.com/aws-amplify/amplify-ui/commit/bb141a719fd9bc2d7680e539f2ff047deb88ee7e)]:
+  - @aws-amplify/ui-react@6.0.3
+  - @aws-amplify/ui@6.0.2
+  - @aws-amplify/ui-react-core@3.0.2
+
+## 3.0.3
+
+### Patch Changes
+
+- Updated dependencies [[`9dd0e58e5`](https://github.com/aws-amplify/amplify-ui/commit/9dd0e58e5167d307c2154b3280de3c0e52f607e1)]:
+  - @aws-amplify/ui@6.0.1
+  - @aws-amplify/ui-react@6.0.2
+  - @aws-amplify/ui-react-core@3.0.1
+
+## 3.0.2
+
+### Patch Changes
+
+- [#4760](https://github.com/aws-amplify/amplify-ui/pull/4760) [`74f834422`](https://github.com/aws-amplify/amplify-ui/commit/74f83442276c73af60da740acf4b7bc2b2878544) Thanks [@thaddmt](https://github.com/thaddmt)! - fix(storagemanager): upload action now displays state properly (#4666)
+
 ## 3.0.1
 
 ### Patch Changes